requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==5.2.0
certifi==2022.12.7
chardet==4.0.0
cleantext==1.1.4
click==8.1.3
exceptiongroup==1.1.0
google==3.0.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlesearch-python==1.1.0
h11==0.14.0
idna==2.10
joblib==1.2.0
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
outcome==1.2.0
pandas==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypdf2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==5.2.0
certifi==2022.12.7
chardet==4.0.0
cleantext==1.1.4
click==8.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
google==3.0.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlesearch-python==1.1.0
h11==0.14.0
idna==2.10
joblib==1.2.0
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
outcome==1.2.0
pandas==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypdf2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
